requirements.txt
tensorflow==2.14.0
numpy==1.24.3
Pillow==9.5.0
scikit-learn==1.3.0
